:sparkles: db query consolidation
