+++ .github/copilot-instructions.md
@@ -1,0 +1,17 @@
+# Core Requirements
+
+- The end golas is stability, speed and great user experience.
+
+## Code Quality Requirements
+
+- Follow standard TypeScript and JavaScript conventions and best practices.
+- Use the Composition API when creating Vue components.
+- Use clear, descriptive variable names and functions.
+- Accessibility should always be a first-class consideration and should be part of the intial planning and design.
+- Add comments to explain complex logic or non-obvious implementations
+- Keep functions focused and manageable (generally under 50 lines)
+- Use error handling patterns consistently
+- This project uses Nuxt v4 directory structure, meaning that `app/` is the source folder and composabled, plugins, pages and so on live inside it.
+- Use `tailwindcss` for styling.
+- Add comments to explain complex logic or non-obvious implementations
+- Ensure you write strictly type-safe code, for example by ensuring you always check when accessing and array value by index.
